requirements.txt
bilibili-api-python
curl_cffi~=0.9.0b1
